Moves new scripts to the misc directory to prevent the root js dir from becoming overly busy.
